RF-929: Move email form into own component
