+++ screens/index.ts
@@ -5,7 +5,7 @@ export { default as BankingDisclosuresScreen } from './BankingDisclosuresScreen'
 export { default as ConfirmPIIScreen } from './ConfirmPIIScreen';
 export { default as DisclosuresScreen } from './DisclosuresScreen';
 export { default as ExternalAccountScreen } from './ExternalAccountScreen';
-export { default as ForgotPasswordScreen } from './ForgotPasswordScreen';
+export { default as ForgotPasswordScreen } from './ForgotPasswordScreen/ForgotPasswordScreen';
 export { default as InitTransferScreen } from './InitTransferScreen';
 export { default as LoginScreen } from './LoginScreen/LoginScreen';
 export { default as MenuScreen } from './MenuScreen';
